[TEMP COMMIT]. Core module update |NOT TESTED|

+++ composeApp/core/domain/src/commonMain/kotlin/com/singing/app/domain/model/Account.kt
@@ -9,6 +9,6 @@ data class UserData(
 )
 
 data class UserInfo(
-    val publicationsCount: Int,
+    val publicationsCount: Long,
     val registeredAt: StableInstant,
 )
