+++ composeApp/core/domain/src/commonMain/kotlin/com/singing/app/domain/model/PublicationTag.kt
@@ -2,5 +2,5 @@ package com.singing.app.domain.model
 
 data class PublicationTagStatistics(
     val title: String,
-    val publications: Int,
+    val publications: Long,
 )
